pyproject.toml
[tool.poetry]
name = "mysqldump-analyzer"
version = "0.1.0"
description = "MYSQL dump analyzer helps you knowing the different between your MYSQL dump files."
authors = ["Back Market contributors"]
keywords = ["mysql", "dump", "analyzer"]
include = ["mysqldump_analyzer/py.typed"]
packages = [
    {include = "mysqldump_analyzer"}
]
license = "LICENSE.txt"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3 :: Only"
]

[tool.poetry.dependencies]
python = "^3.7"
click = "^8.0"
pyyaml = "^6.0"
rich = "^13.3.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.0"
black = "^23.3.0"
mypy = "^1.2"
isort = "^5.11.5"
types-PyYAML = "^6.0"
types-click = "^7.1.8"

[tool.poetry.scripts]
mysqldump_analyzer = 'mysqldump_analyzer:main'

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py38']

[tool.isort]
profile = "black"

[tool.mypy]
files = "./mysqldump_analyzer/**/*.py"
exclude = "^tests"
mypy_path = "./stubs"
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = false
implicit_reexport = false
strict_equality = true
strict_optional = false
warn_unreachable = true
warn_unused_configs = true
